Move and split layout interfaces into layout directory
