--- CourageScores/ClientApp/src/components/tournaments/layout.ts
@@ -1,37 +1,0 @@
-import {GameMatchOptionDto} from "../../interfaces/models/dtos/Game/GameMatchOptionDto";
-import {TournamentMatchDto} from "../../interfaces/models/dtos/Game/TournamentMatchDto";
-import {TournamentSideDto} from "../../interfaces/models/dtos/Game/TournamentSideDto";
-import {TournamentRoundDto} from "../../interfaces/models/dtos/Game/TournamentRoundDto";
-
-export interface ILayoutDataForSide {
-    id: string;
-    name: string;
-    link: JSX.Element;
-    mnemonic?: string;
-    showMnemonic?: boolean;
-}
-
-export interface ILayoutDataForMatch {
-    sideA: ILayoutDataForSide;
-    sideB: ILayoutDataForSide;
-    scoreA: string;
-    scoreB: string;
-    winner?: string;
-    saygId?: string;
-    mnemonic?: string;
-    hideMnemonic?: boolean;
-    matchOptions?: GameMatchOptionDto;
-    match?: TournamentMatchDto;
-    numberOfSidesOnTheNight?: string;
-}
-
-export interface ILayoutDataForRound {
-    name: string;
-    matches: ILayoutDataForMatch[];
-    possibleSides: TournamentSideDto[];
-    alreadySelectedSides: TournamentSideDto[];
-    round?: TournamentRoundDto;
-    preRound?: boolean;
-}
-
-
