+++ CourageScores/ClientApp/src/components/tournaments/layout/ILayoutDataForSide.ts
@@ -1,0 +1,7 @@
+export interface ILayoutDataForSide {
+    id: string;
+    name: string;
+    link: JSX.Element;
+    mnemonic?: string;
+    showMnemonic?: boolean;
+}
