+++ CourageScores/ClientApp/src/components/tournaments/layout/ILayoutEngine.ts
@@ -1,5 +1,5 @@
-import {ILayoutDataForRound} from "../layout";
 import {ILayoutRequest} from "./ILayoutRequest";
+import {ILayoutDataForRound} from "./ILayoutDataForRound";
 
 export interface ILayoutEngine {
     calculate(request: ILayoutRequest): ILayoutDataForRound[]
